\name{powerindexR-package}
\alias{powerindexR-package}
\alias{powerindexR}
\docType{package}
\title{
\packageTitle{powerindexR}
}
\description{
\packageDescription{powerindexR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{powerindexR}
\packageIndices{powerindexR}
This R package allows the determination of some distributions of the voters' 
power when passing laws in weighted voting situations.
}
\author{
Livino M. Armijos-Toro, Jose M. Alonso-Meijide, Manuel A. Mosquera, Alejandro Saavedra-Nieves.
}
\references{
Alonso-Meijide, J. M., & Bowles, C. (2005). Generating functions for coalitional power indices: An application to the IMF. Annals of Operations Research, 137, 21-44. <https://doi.org/10.1007/s10479-005-2242-y>.

Brams, S. J., & Affuso, P. J. (1976). Power and size: A new paradox. Theory and Decision, 7(1-2), 29-56. <https://doi.org/10.1007/BF00141101>.

Colomer, J. M., & Martinez, F. (1995). The paradox of coalition trading. Journal of Theoretical Politics, 7(1), 41-63. <https://doi.org/10.1177/0951692895007001003>.

Johnston, R. J. (1978). On the measurement of power: Some reactions to Laver. Environment and Planning A, 10(8), 907-914. <https://doi.org/10.1068/a100907>.

Lucas, W. F. (1983). Measuring power in weighted voting systems (pp. 183-238). Springer New York. <https://doi.org/10.1007/978-1-4612-5430-0_9>
}

\keyword{powerindex}

